Distributions over Partitions

The following distributions are implemented.

Missing docstring.

Missing docstring for AbstractMvUrnDistribution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UniformMvUrnDistribution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BetaBinomialMvUrnDistribution. Check Documenter's build log for details.

EqualitySampler.CustomInclusionMvUrnDistribution — Type

CustomInclusionMvUrnDistribution is similar to the BetaBinomialMvUrnDistribution in that the model probabilities are completely determined by the size of the partition. Whereas the BetaBinomialMvUrnDistribution uses a BetaBinomial distribution to obtain the probabilities, the CustomInclusionMvUrnDistribution can be used to specify any vector of probabilities. This distribution is particularly useful to sample uniformly from partitions of a given size. For example:

rand(CustomInclusionMvUrnDistribution(4, ntuple(i->log(i==1), Val(4)))) # always all equal (1 parameter)
rand(CustomInclusionMvUrnDistribution(4, ntuple(i->log(i==3), Val(4)))) # always 3 parameters
rand(CustomInclusionMvUrnDistribution(4, ntuple(i->log(i==4), Val(4)))) # always completely distinct (4 parameters)

The function does not check if sum(exp, logpdf) ≈ 1.0, that is the callers responsibility.

Missing docstring.

Missing docstring for RandomProcessMvUrnDistribution. Check Documenter's build log for details.

Aside from the interface for multivariate distributions, the following methods are also defined.

Missing docstring.

Missing docstring for pdf_model, logpdf_model. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pdf_model_distinct. Check Documenter's build log for details.

Missing docstring.

Missing docstring for logpdf_model_distinct. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pdf_incl. Check Documenter's build log for details.

Missing docstring.

Missing docstring for logpdf_incl. Check Documenter's build log for details.